Folder structure and solution generation fixed
Solution now generates properly in VS, and the folder structure for the plugin source is reorganized accordingly.
