--- Unreal_Plugin/UnrealPluginProject/Binaries/Win64/UE4Editor.modules
@@ -1,7 +1,0 @@
-{
-	"BuildId": "3944462",
-	"Modules": 
-	{
-		"UnrealPluginProject": "UE4Editor-UnrealPluginProject.dll"
-	}
-}
