--- Unreal_Plugin/UnrealPluginProject/UnrealPluginProject.sln
@@ -1,72 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.22310.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Engine", "Engine", "{F79FAB3B-70A4-44C3-995C-8BA9103D5F23}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Enterprise", "Enterprise", "{768CA583-A5AD-4D8A-B553-A6B9B51CCB3B}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Games", "Games", "{3F50C74A-AFE8-4C86-9C96-4FD62FFA4274}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UE4", "Intermediate\ProjectFiles\UE4.vcxproj", "{098D36F7-75DC-42E9-A7A6-47EA2CF9961F}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Studio", "Intermediate\ProjectFiles\Studio.vcxproj", "{C1A2F4FA-5AD2-4CCE-A684-0BA06CF5E5B2}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UnrealPluginProject", "Intermediate\ProjectFiles\UnrealPluginProject.vcxproj", "{A2D4EC10-9D88-47EC-99F0-50DC2444874C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		DebugGame Editor|HTML5 = DebugGame Editor|HTML5
-		DebugGame Editor|Win32 = DebugGame Editor|Win32
-		DebugGame Editor|Win64 = DebugGame Editor|Win64
-		DebugGame|HTML5 = DebugGame|HTML5
-		DebugGame|Win32 = DebugGame|Win32
-		DebugGame|Win64 = DebugGame|Win64
-		Development Editor|HTML5 = Development Editor|HTML5
-		Development Editor|Win32 = Development Editor|Win32
-		Development Editor|Win64 = Development Editor|Win64
-		Development|HTML5 = Development|HTML5
-		Development|Win32 = Development|Win32
-		Development|Win64 = Development|Win64
-		Shipping|HTML5 = Shipping|HTML5
-		Shipping|Win32 = Shipping|Win32
-		Shipping|Win64 = Shipping|Win64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame Editor|HTML5.ActiveCfg = HTML5_DebugGame_Editor|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame Editor|Win32.ActiveCfg = DebugGame_Editor|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame Editor|Win64.ActiveCfg = DebugGame_Editor|x64
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame Editor|Win64.Build.0 = DebugGame_Editor|x64
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame|HTML5.ActiveCfg = HTML5_DebugGame_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame|HTML5.Build.0 = HTML5_DebugGame_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame|Win32.ActiveCfg = DebugGame_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame|Win32.Build.0 = DebugGame_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame|Win64.ActiveCfg = DebugGame_Game|x64
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.DebugGame|Win64.Build.0 = DebugGame_Game|x64
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development Editor|HTML5.ActiveCfg = HTML5_Development_Editor|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development Editor|Win32.ActiveCfg = Development_Editor|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development Editor|Win64.ActiveCfg = Development_Editor|x64
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development Editor|Win64.Build.0 = Development_Editor|x64
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development|HTML5.ActiveCfg = HTML5_Development_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development|HTML5.Build.0 = HTML5_Development_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development|Win32.ActiveCfg = Development_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development|Win32.Build.0 = Development_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development|Win64.ActiveCfg = Development_Game|x64
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Development|Win64.Build.0 = Development_Game|x64
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Shipping|HTML5.ActiveCfg = HTML5_Shipping_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Shipping|HTML5.Build.0 = HTML5_Shipping_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Shipping|Win32.ActiveCfg = Shipping_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Shipping|Win32.Build.0 = Shipping_Game|Win32
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Shipping|Win64.ActiveCfg = Shipping_Game|x64
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C}.Shipping|Win64.Build.0 = Shipping_Game|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{098D36F7-75DC-42E9-A7A6-47EA2CF9961F} = {F79FAB3B-70A4-44C3-995C-8BA9103D5F23}
-		{C1A2F4FA-5AD2-4CCE-A684-0BA06CF5E5B2} = {768CA583-A5AD-4D8A-B553-A6B9B51CCB3B}
-		{A2D4EC10-9D88-47EC-99F0-50DC2444874C} = {3F50C74A-AFE8-4C86-9C96-4FD62FFA4274}
-	EndGlobalSection
-EndGlobal
