+++ Unreal_Plugin/UnrealToolkitProject/Binaries/Win64/UE4Editor.modules
@@ -1,0 +1,7 @@
+{
+	"BuildId": "3944462",
+	"Modules": 
+	{
+		"UnrealToolkitProject": "UE4Editor-UnrealToolkitProject.dll"
+	}
+}
