+++ Unreal_Plugin/UnrealToolkitProject/Plugins/BHoM/Source/Unreal_Toolkit/Private/BHoMFunctionLibrary.cpp
@@ -20,6 +20,8 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
+#pragma once
+
 #include "Unreal_Toolkit.h"
 #include "BHoMFunctionLibrary.h"
 #include <list>
