+++ Unreal_Plugin/UnrealToolkitProject/Source/UnrealToolkitProject.Target.cs
@@ -1,0 +1,14 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+using UnrealBuildTool;
+using System.Collections.Generic;
+
+public class UnrealToolkitProjectTarget : TargetRules
+{
+	public UnrealToolkitProjectTarget(TargetInfo Target) : base(Target)
+	{
+		Type = TargetType.Game;
+
+		ExtraModuleNames.AddRange( new string[] { "UnrealToolkitProject" } );
+	}
+}
