+++ Unreal_Plugin/UnrealToolkitProject/Source/UnrealToolkitProject/UnrealToolkitProject.cpp
@@ -1,0 +1,6 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "UnrealToolkitProject.h"
+#include "Modules/ModuleManager.h"
+
+IMPLEMENT_PRIMARY_GAME_MODULE( FDefaultGameModuleImpl, UnrealToolkitProject, "UnrealToolkitProject" );
