+++ Unreal_Plugin/UnrealToolkitProject/Source/UnrealToolkitProject/UnrealToolkitProjectGameModeBase.cpp
@@ -1,0 +1,7 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "UnrealToolkitProjectGameModeBase.h"
+
+
+
+
