+++ Unreal_Plugin/UnrealToolkitProject/Source/UnrealToolkitProject/UnrealToolkitProjectGameModeBase.h
@@ -1,0 +1,20 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/GameModeBase.h"
+#include "UnrealToolkitProjectGameModeBase.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class UNREALTOOLKITPROJECT_API AUnrealToolkitProjectGameModeBase : public AGameModeBase
+{
+	GENERATED_BODY()
+	
+	
+	
+	
+};
