+++ Unreal_Plugin/UnrealToolkitProject/Source/UnrealToolkitProjectEditor.Target.cs
@@ -1,0 +1,14 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+using UnrealBuildTool;
+using System.Collections.Generic;
+
+public class UnrealToolkitProjectEditorTarget : TargetRules
+{
+	public UnrealToolkitProjectEditorTarget(TargetInfo Target) : base(Target)
+	{
+		Type = TargetType.Editor;
+
+		ExtraModuleNames.AddRange( new string[] { "UnrealToolkitProject" } );
+	}
+}
